Hotfix/pipelinescript fixes (#679)
- Parameter and Logging Enhancements: Added FORCE_RESET parameter and enhanced debug logging in control plane deployment pipeline. Also, added warning logs for forced re-install in control plane preparation script.
- Script Improvements: Refactored various scripts for consistency, readability, and error handling. This includes adding shebangs, fixing typos, and improving output formatting.
- Key Vault and Terraform Updates: Updated Key Vault configurations, enhanced logging for remote Terraform state management, and added dependencies for proper sequencing and resource creation.
- Resource Management: Removed unnecessary dependencies and variables, such as DEPLOYER_RANDOM_ID and LIBRARY_RANDOM_ID, and improved handling of custom random IDs.
- Error Handling and Output Management: Improved error handling and output management across multiple scripts, ensuring proper return value assignments and clear messaging for Terraform commands.

---------

+++ deploy/scripts/helpers/script_helpers.sh
@@ -538,13 +538,15 @@ function ImportAndReRunApply {
 						return_value=$?
 						echo "Error when importing resource"
 						echo "Terraform import:                      failed"
-
+						if [ -f "$fileName" ]; then
+							rm "$fileName"
+						fi
 						return $return_value
 					else
 						echo "Terraform import:                      succeeded"
 					fi
 				done
-				rm "$fileName"
+
         # shellcheck disable=SC2086
 				if ! terraform -chdir="${terraform_module_directory}" plan -input=false $allImportParameters ; then
 						echo ""
